address comments, test updated
